// src/api/openapi/model/offerTemplateEntry.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OfferElement } from './offerElement';


export interface OfferTemplateEntry { 
    children?: Array<OfferTemplateEntry>;
    element: OfferElement;
    id: number;
}