// src/api/openapi/model/offerTemplateEntryCreatePatch.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OfferTemplateEntryCreatePatch { 
    children?: Array<OfferTemplateEntryCreatePatch>;
    elementId: number;
}